+++ components/Layout/Nav.jsx
@@ -75,7 +75,13 @@ const UserMenu = ({ user, mutate }) => {
               <a className={styles.item}>Profile</a>
             </Link>
             <Link passHref href="/settings">
-              <a className={styles.item}>Settngs</a>
+              <a className={styles.item}>Settings</a>
+            </Link>
+            <Link passHref href="/blog">
+              <a className={styles.item}>Blog</a>
+            </Link>
+            <Link passHref href="/post">
+              <a className={styles.item}>Manage Posts</a>
             </Link>
             <div className={styles.item} style={{ cursor: 'auto' }}>
               <Container alignItems="center">