+++ page-components/Blog/index.jsx
@@ -1,13 +1,11 @@
 import { Spacer } from '@/components/Layout';
-import styles from './Feed.module.css';
-import Poster from './Poster';
+import styles from './styles.module.css';
 import PostList from './PostList';
 
-export const Feed = () => {
+export const Post = () => {
   return (
     <div className={styles.root}>
       <Spacer size={1} axis="vertical" />
-      <Poster />
       <PostList />
     </div>
   );