+++ pages/blog.jsx
@@ -1,0 +1,15 @@
+import { Post } from '@/page-components/Blog';
+import Head from 'next/head';
+
+const PostPage = () => {
+  return (
+    <>
+      <Head>
+        <title>Blog</title>
+      </Head>
+      <Post />
+    </>
+  );
+};
+
+export default PostPage;